Remove commented code from get-latest.sh
This code seems to be in preparation for a Tomcat Native Spec File that
was never implemented.

A variable name was changed in such a way as to allow the adding back of
the Tomcat Native code if and when it is ready, while still making sense in
its present state.