Append major version number
The package name will now include the major version number of that
particular Tomcat version. This helps with testing on Fedora and other
distros which already contain a package called tomcat. Without this
change, RPM might try to update the package to a repository version of
it.

The RPM spec file now recommends packages suitable for use with Tomcat
and Axelor ERP. It should be compatible with all RHEL-based distros, as
well as openSUSE.

A regression from commit 1d9e0bf92a415cfd909243e19cd2789e4e1bec72 was
fixed.

Fixes #3